requirements.txt
selenium
python-dotenv
speechrecognition
pyaudio
pylint